--- validate-lib/src/main/java/no/difi/vefa/message/MessageType.java
@@ -1,9 +1,0 @@
-package no.difi.vefa.message;
-
-/**
- * This enum is used to describe different message types.
- */
-public enum MessageType {
-	Warning,
-	Fatal
-}
