--- validate-lib/src/main/java/no/difi/vefa/message/ValidationType.java
@@ -1,12 +1,0 @@
-package no.difi.vefa.message;
-
-/**
- * This enum is used to describe different validation types.
- */
-public enum ValidationType {
-	XMLWellFormed,
-	XSD,
-	XSL,
-	Filter,
-	Configuration
-}
