+++ validate-lib/src/main/java/no/difi/vefa/model/message/Hint.java
@@ -1,0 +1,32 @@
+package no.difi.vefa.model.message;
+
+import javax.xml.bind.annotation.XmlAccessType;
+import javax.xml.bind.annotation.XmlAccessorType;
+import javax.xml.bind.annotation.XmlElement;
+import javax.xml.bind.annotation.XmlRootElement;
+
+@XmlRootElement(name = "hint")
+@XmlAccessorType(XmlAccessType.PROPERTY)
+public class Hint {
+
+	private String title;
+    private String description;
+
+    @XmlElement(name = "title")
+    public String getTitle() {
+        return title;
+    }
+
+    public void setTitle(String title) {
+        this.title = title;
+    }
+
+    @XmlElement(name = "description")
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+}
