+++ validate-lib/src/main/java/no/difi/vefa/model/message/MessageType.java
@@ -1,0 +1,9 @@
+package no.difi.vefa.model.message;
+
+/**
+ * This enum is used to describe different message types.
+ */
+public enum MessageType {
+	Warning,
+	Fatal
+}
