+++ validate-lib/src/main/java/no/difi/vefa/model/message/Messages.java
@@ -1,4 +1,4 @@
-package no.difi.vefa.message;
+package no.difi.vefa.model.message;
 
 import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlRootElement;
