+++ validate-lib/src/main/java/no/difi/vefa/model/message/ValidationType.java
@@ -1,0 +1,12 @@
+package no.difi.vefa.model.message;
+
+/**
+ * This enum is used to describe different validation types.
+ */
+public enum ValidationType {
+	XMLWellFormed,
+	XSD,
+	XSL,
+	Filter,
+	Configuration
+}
