+++ validate-lib/src/main/java/no/difi/vefa/validation/RenderXsl.java
@@ -1,8 +1,8 @@
 package no.difi.vefa.validation;
 
-import no.difi.vefa.message.Messages;
-import no.difi.vefa.util.MessageUtils;
-import no.difi.vefa.util.xml.XmlXslTransformation;
+import no.difi.vefa.model.message.Messages;
+import no.difi.vefa.utils.MessageUtils;
+import no.difi.vefa.utils.xml.XmlXslTransformation;
 import org.w3c.dom.Document;
 
 /**
