+++ validate-lib/src/test/java/no/difi/vefa/validation/SchematronTransformationTest.java
@@ -1,9 +1,9 @@
 package no.difi.vefa.validation;
 
-import no.difi.vefa.message.MessageType;
-import no.difi.vefa.message.Messages;
-import no.difi.vefa.message.ValidationType;
-import no.difi.vefa.util.xml.XmlUtils;
+import no.difi.vefa.model.message.MessageType;
+import no.difi.vefa.model.message.Messages;
+import no.difi.vefa.model.message.ValidationType;
+import no.difi.vefa.utils.xml.XmlUtils;
 import org.junit.Before;
 import org.junit.Test;
 import org.w3c.dom.Document;
