+++ validate-lib/src/test/java/no/difi/vefa/validation/ValidateTest.java
@@ -1,12 +1,12 @@
 package no.difi.vefa.validation;
 
-import no.difi.vefa.message.Hint;
-import no.difi.vefa.message.Message;
-import no.difi.vefa.message.MessageType;
-import no.difi.vefa.message.Messages;
-import no.difi.vefa.message.ValidationType;
-import no.difi.vefa.util.PropertiesUtils;
-import no.difi.vefa.util.xml.XmlUtils;
+import no.difi.vefa.model.message.Hint;
+import no.difi.vefa.model.message.Message;
+import no.difi.vefa.model.message.MessageType;
+import no.difi.vefa.model.message.Messages;
+import no.difi.vefa.model.message.ValidationType;
+import no.difi.vefa.utils.PropertiesUtils;
+import no.difi.vefa.utils.xml.XmlUtils;
 import org.junit.Before;
 import org.junit.Test;
 import org.w3c.dom.Document;
