+++ validate-ws/src/main/java/no/difi/vefa/ws/rest/model/ListVersions.java
@@ -1,8 +1,8 @@
 package no.difi.vefa.ws.rest.model;
 
-import no.difi.vefa.configuration.Configuration;
-import no.difi.vefa.util.PropertiesUtils;
-import no.difi.vefa.util.xml.XmlUtils;
+import no.difi.vefa.utils.configuration.ConfigurationUtils;
+import no.difi.vefa.utils.PropertiesUtils;
+import no.difi.vefa.utils.xml.XmlUtils;
 import org.w3c.dom.Document;
 import org.w3c.dom.NodeList;
 
@@ -26,11 +26,11 @@ public class ListVersions {
      * @throws Exception
      */
     public String getVersions() throws Exception {
-        Configuration configuration = new Configuration();
+        ConfigurationUtils configurationUtils = new ConfigurationUtils();
         ArrayList<String> list = new ArrayList<>();
-        Document standardConfig = configuration.fileToXMLDOM(propertiesUtils.dataDir + "/STANDARD/config.xml", propertiesUtils);
+        Document standardConfig = configurationUtils.fileToXMLDOM(propertiesUtils.dataDir + "/STANDARD/config.xml", propertiesUtils);
         this.addVersionToList(standardConfig, list);
-        Document customConfig = configuration.fileToXMLDOM(propertiesUtils.dataDir + "/CUSTOM/config.xml", propertiesUtils);
+        Document customConfig = configurationUtils.fileToXMLDOM(propertiesUtils.dataDir + "/CUSTOM/config.xml", propertiesUtils);
         this.addVersionToList(customConfig, list);
         Set<String> set = new HashSet<>(list);
         List<String> uniqueList = new ArrayList<>(set);
